:bug: fix: add Lifespan._on_ready() for forward adapter startup
fix #2475
